src/main/java/com/example/demo/repository/MemoryMemberRepository.java
package com.example.demo.repository;

import com.example.demo.domain.Member;
import org.springframework.stereotype.Repository;

import java.util.*;

public class MemoryMemberRepository implements MemberRepository {

    private static Map<Long,Member> store = new HashMap<>();
    private static Long sequence = 0L;

    @Override
    public Member save(Member member) {
        member.setId(++sequence);
        store.put(member.getId(),member);
        return null;
    }

    @Override
    public Optional<Member> findById(Long id) {
        return Optional.ofNullable(store.get(id));
    }

    @Override
    public Optional<Member> findByName(String name) {
        //stream 객체 생성
        return store.values().stream()
                // name과 동일한 멤버 정보를 찾는다
                .filter(member -> member.getName().equals(name))
                // 일치하는 요소 한개 리턴
                .findAny();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore(){
        //데이터 초기화
        store.clear();
    }
}
